Converted ITGIO to uint16_t, which it should be according to documentation
git-svn-id: https://openitg.svn.sourceforge.net/svnroot/openitg@170 83fadc84-e282-4d84-a09a-c4228d6ae7e5

+++ src/io/ITGIO.h
@@ -11,8 +11,8 @@ public:
 //	all in one place. It's not important right now, though. - Vyhd
 //	static bool DeviceMatches( int idVendor, int idProduct );
 
-	bool Read( uint32_t *pData );
-	bool Write( uint32_t iData );
+	bool Read( uint16_t *pData );
+	bool Write( uint16_t iData );
 
 	/* Globally accessible. */
 	static int m_iInputErrorCount;
